+++ functions.php
@@ -394,6 +394,28 @@ if (!function_exists('narga_comment_pagination')) :
 endif;
 
 /**
+ * Generator breadcrumbs with Foundation
+ *
+ * Since NARGA v1.3.5
+ *
+ **/
+function narga_breadcrumb() {
+    if (!is_home()) {
+        echo '<ul class="breadcrumbs"><li><a href="';
+        echo home_url();
+        echo '">';
+        bloginfo('name');
+        echo "</a></li>";
+        if (is_category()) {
+            the_category('<li>', '</li>');
+        } elseif (is_page() || is_single()) {
+            echo the_title('<li class="current">', '</li>');
+        } 
+        echo "</ul>";
+    }
+}
+
+/**
  * Add Framework Customizer Direct Link
  *
  * Since NARGA v1.1